% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejsScatter.R
\name{renderThreejsScatter}
\alias{renderThreejsScatter}
\title{Shiny render for threejsScatterOutput}
\usage{
renderThreejsScatter(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a threejsScatter}

\item{env}{The environment in which to evaluate \code{expr}.}
}
\description{
Shiny render for threejsScatterOutput
}
