% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejsScatter.R
\name{threejsScatterOutput}
\alias{threejsScatterOutput}
\title{Shiny output bindings for a threejsScatter}
\usage{
threejsScatterOutput(outputId, width = "100\%", height = "800px")
}
\arguments{
\item{outputId}{output id of the widget}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}
}
\description{
Shiny output bindings for a threejsScatter
}
