% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejsScatter.R
\name{threejsScatter}
\alias{threejsScatter}
\title{threejsScatter}
\usage{
threejsScatter(data, dimX, dimY, dimZ, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{a list of options and data points to pass to the scatter plot}

\item{dimX, dimY, dimZ}{how long you want the x, y, and z dimensions of the graph}
}
\value{
html widget
}
\description{
Creates a 3D scatter plot html widget using threejs.
}
